Settings Integration

+++ java/capstone/wumaps/MainActivity.java
@@ -55,7 +55,7 @@ public class MainActivity extends AppCompatActivity {
 
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView((int) R.layout.activity_main);
+        setContentView(R.layout.activity_main);
         this.addClassButton = (Button) findViewById(R.id.addClassButton);
         this.addClassButton.setOnClickListener(new MyListener());
         this.buildingSearchButton = (Button) findViewById(R.id.buildingSearchButton);
